Move vertical scrollbar to components
