+++ src/screens/resources/podcasts-desktop.js
@@ -1,4 +1,4 @@
-import { VerticalScrollbar } from "../videos/vertical-scrollbar";
+import { VerticalScrollbar } from "../../components/vertical-scrollbar";
 import { podcasts } from "./resources";
 
 export function PodcastsDesktop({ className }) {
@@ -32,7 +32,7 @@ export function PodcastsDesktop({ className }) {
 
             <div className="mt-auto flex gap-2">
               <a
-                className="rounded-full bg-brown-lighter px-5 py-1 font-bold uppercase text-white"
+                className="bg-brown-lighter rounded-full px-5 py-1 font-bold uppercase text-white"
                 href={podcast.listenUrl}
                 target="_blank"
                 rel="noreferrer"
@@ -40,7 +40,7 @@ export function PodcastsDesktop({ className }) {
                 Listen now
               </a>
               <a
-                className="rounded-full bg-brown px-5 py-1 font-bold uppercase text-white"
+                className="bg-brown rounded-full px-5 py-1 font-bold uppercase text-white"
                 href={podcast.websiteUrl}
                 target="_blank"
                 rel="noreferrer"
@@ -48,7 +48,7 @@ export function PodcastsDesktop({ className }) {
                 Website
               </a>
               <a
-                className="rounded-full bg-brown px-5 py-1 font-bold uppercase text-white"
+                className="bg-brown rounded-full px-5 py-1 font-bold uppercase text-white"
                 href={podcast.donateUrl}
                 target="_blank"
                 rel="noreferrer"
